+++ packages/facility-server/__tests__/apiv1/surveyResponse/helpers.js
@@ -1,0 +1,201 @@
+import { fake } from '@tamanu/fake-data/fake';
+import { PROGRAM_DATA_ELEMENT_TYPES, SURVEY_TYPES } from '@tamanu/constants';
+import { getCurrentDateTimeString } from '@tamanu/utils/dateTime';
+
+export const buildPatchBody = overrides => ({ editedTime: getCurrentDateTimeString(), ...overrides });
+
+export const createSurveyResponseTestHelpers = models => {
+  const setupAutocompleteSurvey = async (sscConfig, answerBody) => {
+    const {
+      Facility,
+      Location,
+      Department,
+      Patient,
+      User,
+      Encounter,
+      Program,
+      Survey,
+      SurveyResponse,
+      ProgramDataElement,
+      SurveyScreenComponent,
+      SurveyResponseAnswer,
+    } = models;
+
+    const facility = await Facility.create(fake(Facility));
+    const location = await Location.create({
+      ...fake(Location),
+      facilityId: facility.id,
+    });
+    const department = await Department.create({
+      ...fake(Department),
+      facilityId: facility.id,
+    });
+    const examiner = await User.create(fake(User));
+    const patient = await Patient.create(fake(Patient));
+    const encounter = await Encounter.create({
+      ...fake(Encounter),
+      patientId: patient.id,
+      departmentId: department.id,
+      locationId: location.id,
+      examinerId: examiner.id,
+    });
+    const program = await Program.create(fake(Program));
+    const survey = await Survey.create({
+      ...fake(Survey),
+      programId: program.id,
+    });
+    const response = await SurveyResponse.create({
+      ...fake(SurveyResponse),
+      surveyId: survey.id,
+      encounterId: encounter.id,
+    });
+    const dataElement = await ProgramDataElement.create({
+      ...fake(ProgramDataElement),
+      type: 'Autocomplete',
+    });
+    await SurveyScreenComponent.create({
+      ...fake(SurveyScreenComponent),
+      responseId: response.id,
+      dataElementId: dataElement.id,
+      surveyId: survey.id,
+      config: sscConfig,
+    });
+    const answer = await SurveyResponseAnswer.create({
+      ...fake(SurveyResponseAnswer),
+      dataElementId: dataElement.id,
+      responseId: response.id,
+      body: answerBody,
+    });
+
+    return { answer, response, facilityId: facility.id };
+  };
+
+  const setupAutocompleteSurveyWithoutAnswer = async sscConfig => {
+    const {
+      Facility,
+      Location,
+      Department,
+      Patient,
+      User,
+      Encounter,
+      Program,
+      Survey,
+      SurveyResponse,
+      ProgramDataElement,
+      SurveyScreenComponent,
+    } = models;
+
+    const facility = await Facility.create(fake(Facility));
+    const location = await Location.create({
+      ...fake(Location),
+      facilityId: facility.id,
+    });
+    const department = await Department.create({
+      ...fake(Department),
+      facilityId: facility.id,
+    });
+    const examiner = await User.create(fake(User));
+    const patient = await Patient.create(fake(Patient));
+    const encounter = await Encounter.create({
+      ...fake(Encounter),
+      patientId: patient.id,
+      departmentId: department.id,
+      locationId: location.id,
+      examinerId: examiner.id,
+    });
+    const program = await Program.create(fake(Program));
+    const survey = await Survey.create({
+      ...fake(Survey),
+      programId: program.id,
+    });
+    const response = await SurveyResponse.create({
+      ...fake(SurveyResponse),
+      surveyId: survey.id,
+      encounterId: encounter.id,
+    });
+    const dataElement = await ProgramDataElement.create({
+      ...fake(ProgramDataElement),
+      type: 'Autocomplete',
+    });
+    await SurveyScreenComponent.create({
+      ...fake(SurveyScreenComponent),
+      responseId: response.id,
+      dataElementId: dataElement.id,
+      surveyId: survey.id,
+      config: sscConfig,
+    });
+
+    return { dataElement, response, facilityId: facility.id };
+  };
+
+  const setupComplexChartSurvey = async () => {
+    const {
+      Facility,
+      Location,
+      Department,
+      Patient,
+      User,
+      Encounter,
+      Program,
+      Survey,
+      SurveyResponse,
+      SurveyScreenComponent,
+      ProgramDataElement,
+      SurveyResponseAnswer,
+    } = models;
+
+    const facility = await Facility.create(fake(Facility));
+    const location = await Location.create({
+      ...fake(Location),
+      facilityId: facility.id,
+    });
+    const department = await Department.create({
+      ...fake(Department),
+      facilityId: facility.id,
+    });
+    const examiner = await User.create(fake(User));
+    const patient = await Patient.create(fake(Patient));
+    const encounter = await Encounter.create({
+      ...fake(Encounter),
+      patientId: patient.id,
+      departmentId: department.id,
+      locationId: location.id,
+      examinerId: examiner.id,
+    });
+    const program = await Program.create(fake(Program));
+    const survey = await Survey.create({
+      ...fake(Survey),
+      surveyType: SURVEY_TYPES.COMPLEX_CHART_CORE,
+      programId: program.id,
+    });
+    const response = await SurveyResponse.create({
+      ...fake(SurveyResponse),
+      surveyId: survey.id,
+      encounterId: encounter.id,
+    });
+    const dataElement = await ProgramDataElement.create({
+      ...fake(ProgramDataElement),
+      type: PROGRAM_DATA_ELEMENT_TYPES.TEXT,
+    });
+    await SurveyScreenComponent.create({
+      ...fake(SurveyScreenComponent),
+      dataElementId: dataElement.id,
+      surveyId: survey.id,
+      calculation: '',
+    });
+    const answer = await SurveyResponseAnswer.create({
+      ...fake(SurveyResponseAnswer),
+      dataElementId: dataElement.id,
+      responseId: response.id,
+      body: 'Initial answer',
+    });
+
+    return { answer, response, dataElement, survey, facilityId: facility.id };
+  };
+
+  return {
+    setupAutocompleteSurvey,
+    setupAutocompleteSurveyWithoutAnswer,
+    setupComplexChartSurvey,
+  };
+};
